requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
transformers>=4.35.0
torch>=2.0.0
pdfplumber>=0.10.0
faiss-cpu>=1.7.4
huggingface-hub>=0.17.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0